converted server.ts file to server.js
