--- src/server/server.ts
@@ -1,22 +1,0 @@
-import express from "express";
-import cors from "cors";
-import { createClient } from '@supabase/supabase-js'
-
-const app = express();
-const PORT = process.env.PORT || 3000;
-
-// Initialize Supabase
-const supabaseUrl = process.env.VITE_SUPABASE_URL
-const supabaseAnonKey = process.env.VITE_SUPABASE_ANON_KEY
-const supabase = createClient(supabaseUrl!, supabaseAnonKey!)
-
-console.log(supabaseUrl, supabaseAnonKey)
-
-// Middleware
-app.use(cors());
-app.use(express.json());
-
-// Start the server
-app.listen(PORT, () => {
-  console.log(`Server is running on http://localhost:${PORT}`);
-});
